Fix clang build for using DR in google3
None of these are bugs in core DR.

- parse_int(): base is unsigned, comparing vs 0 is tautalogical
- atomic_compare_exchange_int64(): the inline asm is invalid on ia32, but
  gcc doesn't complain since it is unused
- init_patch_list(): clang says the comparison vs USHORT_MAX is
  tautalogical based on the range of the enum type.  Casting to int
  silences the warning.
- linux.fib-static: Clang can't link libc with -static and -rdynamic,
  which is in the default CMAKE_SHARED_LIBRARY_LINK_C_FLAGS.  Adding
  -Wl,--no-export-dynamic makes it work.
- linux.vfork run_child(): no return from non-void function

+++ core/x86/emit_utils.c
@@ -2157,7 +2157,7 @@ void
 init_patch_list(patch_list_t *patch, patch_list_type_t type)
 {
     patch->num_relocations = 0;
-    ASSERT_TRUNCATE(patch->type, ushort, type);
+    ASSERT_TRUNCATE(patch->type, ushort, (int)type);
     patch->type = (ushort) type;
 }
